Refactored Errors to single Response Error class
