+++ app/src/main/java/org/listenbrainz/android/model/Error.kt
@@ -1,8 +1,13 @@
 package org.listenbrainz.android.model
 
+import com.google.gson.Gson
+import com.google.gson.reflect.TypeToken
 import retrofit2.Response
 
-enum class Error(override val genericToast: String, override var actualResponse: String? = null) : ResponseError {
+/** These exceptions need to handled in view-model, shown via UI and not just thrown.
+ * @param genericToast Generic message for the error.
+ * @param actualResponse Actual response given by the server.*/
+enum class ResponseError(val genericToast: String, var actualResponse: String? = null) {
     
     DOES_NOT_EXIST(genericToast = "Error! Object not found."),     // "User Some_User_That_Does_Not_Exist not found"
     
@@ -18,19 +23,24 @@ enum class Error(override val genericToast: String, override var actualResponse:
     
     BAD_GATEWAY(genericToast = "Error! Bad gateway."),
     
-    SERVICE_UNAVAILABLE(genericToast = "Server outage detected."),
+    SERVICE_UNAVAILABLE(genericToast = "Server outage detected. Please try again later."),
     
-    NETWORK_ERROR(genericToast = "App is experiencing network issues at the moment."),
+    NETWORK_ERROR(genericToast = "Network issues detected. Make sure device is connected to internet."),
     
     FILE_NOT_FOUND(genericToast = "Selected file does not exist."),
     
     UNKNOWN(genericToast = "Some error has occurred.");
     
+    
+    /** Simple function that returns the most suitable message to show the user.*/
+    fun toast(): String = actualResponse ?: genericToast
+    
+    
     companion object {
         
         /** Get [ResponseError] for a given Retrofit **error** [Response] from server.*/
         fun <T> getError(response: Response<T>) : ResponseError {
-            val apiError = ResponseError.parseError(response)
+            val apiError = parseError(response)
             val error = apiError.error
             val code = apiError.code
             return when (code) {
@@ -45,6 +55,14 @@ enum class Error(override val genericToast: String, override var actualResponse:
                 else -> UNKNOWN
             }.apply { actualResponse = error }
         }
+    
+        /** Parsing server response into [ApiError] class. Consider using specific functions like [getSocialResponseError], etc. for each repository if
+         * returning errors is the sole motive.*/
+        fun <T> parseError(response: Response<T>) : ApiError =
+            Gson().fromJson(
+                /* json = */ response.errorBody()?.string(),
+                /* typeOfT = */ object : TypeToken<ApiError>() {}.type
+            )
         
     }
