+++ app/src/main/java/org/listenbrainz/android/ui/components/ErrorBar.kt
@@ -16,7 +16,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.delay
-import org.listenbrainz.android.model.GeneralError
 import org.listenbrainz.android.model.ResponseError
 import org.listenbrainz.android.ui.theme.ListenBrainzTheme
 
@@ -62,6 +61,6 @@ fun ErrorBar(
 @Composable
 private fun ErrorBarPreview() {
     ListenBrainzTheme {
-        ErrorBar(error = GeneralError.NETWORK_ERROR) {}
+        ErrorBar(error = ResponseError.NETWORK_ERROR) {}
     }
 }
